+++ components/src/lib/Modal/Modal.tsx
@@ -4,7 +4,8 @@ import { ButtonSize, ButtonStyle, ButtonWidth } from '../Button/Button';
 
 import Button from '../Button/Button';
 import CLOSE_BUTTON from '../../asset/close-button.svg';
-import { useEffect } from 'react';
+import useBlockedScroll from '../hooks/useBlockedScroll';
+import useEscKeyDown from '../hooks/useEscKeyDown';
 
 export type ModalPosition = 'center' | 'bottom';
 export type ButtonPosition = 'row' | 'column';
@@ -45,22 +46,8 @@ const Modal = ({
   primaryColor,
   showCloseButton = false,
 }: ModalProps) => {
-  useEffect(() => {
-    if (isOpened) document.body.style.overflow = 'hidden';
-
-    const handleKeyDownEsc = (e: KeyboardEvent) => {
-      if (e.key === 'Escape' && isOpened) {
-        closeModal();
-      }
-    };
-
-    document.body.addEventListener('keydown', handleKeyDownEsc);
-
-    return () => {
-      document.body.style.overflow = 'scroll';
-      document.body.removeEventListener('keydown', handleKeyDownEsc);
-    };
-  }, [isOpened]);
+  useBlockedScroll(isOpened);
+  useEscKeyDown(closeModal);
 
   return (
     <>
