+++ components/src/lib/hooks/useEscKeyDown.ts
@@ -1,0 +1,19 @@
+import { useEffect } from 'react';
+
+const useEscKeyDown = (callback: () => void) => {
+  const handleKeyDownEsc = (e: KeyboardEvent) => {
+    if (e.key === 'Escape') {
+      callback();
+    }
+  };
+
+  useEffect(() => {
+    document.body.addEventListener('keydown', handleKeyDownEsc);
+
+    return () => {
+      document.body.removeEventListener('keydown', handleKeyDownEsc);
+    };
+  }, []);
+};
+
+export default useEscKeyDown;
